Fix odometry by adding missing Y component
The odometry was only computed taking in account linear velocity on the x-axis and angular
velocity around the z-axis. The code is modified to include linear velocity on the y-axis as well
because the robot is holonomic. Odometry arc and radius are calculated the same way as before but
using the combined linear velocity instead. The end position is then rotated by the angle between the
combined velocity vector and the x-axis